Changed constructors to use XMLTreeNode. Added emitter property to renderable objects. Changed renderable objects visible value to true.

+++ Core/Enemies/EasyEnemy.cpp
@@ -34,7 +34,7 @@ CEasyEnemy::CEasyEnemy(std::string mesh, std::string name, Vect3f position,  flo
   //InitWpVector(2, 4);
 }
 //Nuevo sistema
-CEasyEnemy::CEasyEnemy(CRenderableObject *renderableObject, std::vector<Vect3f> wayPoints, float speed, float turnSpeed, float speedAttack, float life, std::string fsmName, Vect2f characterSize,  float AttackDistance, float zone, bool isStatic) :
+CEasyEnemy::CEasyEnemy(CRenderableObject *renderableObject, std::vector<Vect3f> wayPoints, float speed, float turnSpeed, float speedAttack, int life, std::string fsmName, Vect2f characterSize,  float AttackDistance, float zone, bool isStatic) :
   CEnemy(renderableObject, speed, turnSpeed, life, characterSize, AttackDistance, zone),
   m_WpVector(wayPoints),
   m_CurrentWp(NULL),
